+++ pydoe/space_filling/quasi_random/__init__.py
@@ -1,5 +1,6 @@
 from .cranley_patterson_shift import cranley_patterson_shift
 from .halton import halton_sequence
+from .hammersley import hammersley_sequence
 from .korobov import korobov_sequence
 from .rank1 import rank1_lattice
 from .sobol import sobol_sequence
@@ -9,6 +10,7 @@ from .sukharev import sukharev_grid
 __all__ = [
     "cranley_patterson_shift",
     "halton_sequence",
+    "hammersley_sequence",
     "korobov_sequence",
     "rank1_lattice",
     "sobol_sequence",