+++ pydoe/utils/__init__.py
@@ -1,6 +1,12 @@
 from .build_regression_matrix import build_regression_matrix
+from .iman_conover import iman_conover
 from .utils import scale_samples
 from .var_regression_matrix import var_regression_matrix
 
 
-__all__ = ["build_regression_matrix", "scale_samples", "var_regression_matrix"]
+__all__ = [
+    "build_regression_matrix",
+    "iman_conover",
+    "scale_samples",
+    "var_regression_matrix",
+]